+++ change_log.md
@@ -2,6 +2,7 @@
 
 # Version 1.1
 
+* New: See console log in Modnix.
 * Fix: mods_root added as a reserved id. It is used by mod api path action.
 * Fix: Mod api log action will return true when successfully added to queue.
 * Fix: Unknown mod api action will trigger a warning.